+++ example/lib/basic_example.dart
@@ -3,9 +3,18 @@ import 'package:catcher/catcher_plugin.dart';
 
 main() {
   CatcherOptions debugOptions = CatcherOptions(DialogReportMode(), [
-    ConsoleHandler(),
-    HttpHandler(HttpRequestType.post, Uri.parse("https://httpstat.us/200"),
+    EmailManualHandler(
+        ["jhomlala@gmail.com"],
+        enableDeviceParameters: true,
+        enableStackTrace: true,
+        enableCustomParameters: true,
+        enableApplicationParameters: true,
+        sendHtml: true,
+        emailTitle: "Sample Title",
+        emailHeader: "Sample Header",
         printLogs: true)
+    /*HttpHandler(HttpRequestType.post, Uri.parse("https://httpstat.us/200"),
+        printLogs: true)*/
   ]);
   CatcherOptions releaseOptions = CatcherOptions(NotificationReportMode(), [
     EmailManualHandler(["recipient@email.com"])